requirements.txt
aiogram-dialog==2.0.0
aiogram==3.1.1
aiohttp
aiofiles
aiosqlite
cachetools
coloredlogs
pyrogram>=2.0.0
redis
tgcrypto
toml
tortoise-orm
aerich
asyncpg
psycopg2
watchdog
